+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 * Added rate limit model & handles it for password attempt
 * Fixed rights for user update
 * Added required to coordinates in campus model
+* Added rights to drivers-positions.mjs
+* Scoped CRUD rights to campus
 ## Version 1.4.5
 ### Features
 * Added connection status management for drivers through sockets